// typings/GroupedComboboxProps.d.ts
export type SourceEnum = "context" | "database" | "static";

export type OptionsSourceTypeEnum = "association" | "enumeration" | "boolean";

export type OptionsSourceAssociationCaptionTypeEnum = "attribute" | "expression";

export type OptionsSourceDatabaseCaptionTypeEnum = "attribute" | "expression";

export interface OptionsSourceStaticDataSourceType {
    staticDataSourceValue: DynamicValue<string | Big | boolean | Date>;
    staticDataSourceCustomContent: ReactNode;
    staticDataSourceCaption: DynamicValue<string>;
}

export type OptionsSourceAssociationCustomContentTypeEnum = "yes" | "listItem" | "no";

export type OptionsSourceDatabaseCustomContentTypeEnum = "yes" | "listItem" | "no";

export type StaticDataSourceCustomContentTypeEnum = "yes" | "listItem" | "no";

export type SelectionMethodEnum = "checkbox" | "rowclick";

export type SelectedItemsStyleEnum = "text" | "boxes";

export type CustomEditabilityEnum = "default" | "never" | "conditionally";

export type ReadOnlyStyleEnum = "bordered" | "text";

export type LoadingTypeEnum = "spinner" | "skeleton";

export type SelectedItemsSortingEnum = "caption" | "none";

export type FilterTypeEnum = "contains" | "containsExact" | "startsWith" | "none";

export interface OptionsSourceStaticDataSourcePreviewType {
    staticDataSourceValue: string;
    staticDataSourceCustomContent: { widgetCount: number; renderer: ComponentType<{ children: ReactNode; caption?: string }> };
    staticDataSourceCaption: string;
}

export interface GroupedComboboxContainerProps {
    name: string;
    tabIndex?: number;
    id: string;
    source: SourceEnum;
    optionsSourceType: OptionsSourceTypeEnum;
    attributeEnumeration: EditableValue<string>;
    attributeBoolean: EditableValue<boolean>;
    optionsSourceDatabaseDataSource?: ListValue;
    optionsSourceDatabaseItemSelection?: SelectionSingleValue | SelectionMultiValue;
    optionsSourceAssociationCaptionType: OptionsSourceAssociationCaptionTypeEnum;
    optionsSourceDatabaseCaptionType: OptionsSourceDatabaseCaptionTypeEnum;
    optionsSourceAssociationCaptionAttribute?: ListAttributeValue<string>;
    optionsSourceDatabaseCaptionAttribute?: ListAttributeValue<string>;
    optionsSourceAssociationCaptionExpression?: ListExpressionValue<string>;
    optionsSourceDatabaseCaptionExpression?: ListExpressionValue<string>;
    optionsSourceAssociationGroupAttribute?: ListAttributeValue<string>;
    optionsSourceDatabaseGroupAttribute?: ListAttributeValue<string>;
    databaseAttributeString?: EditableValue<string | Big>;
    optionsSourceDatabaseValueAttribute?: ListAttributeValue<string | Big>;
    attributeAssociation: ReferenceValue | ReferenceSetValue;
    optionsSourceAssociationDataSource?: ListValue;
    staticAttribute: EditableValue<string | Big | boolean | Date>;
    optionsSourceStaticDataSource: OptionsSourceStaticDataSourceType[];
    emptyOptionText?: DynamicValue<string>;
    noOptionsText?: DynamicValue<string>;
    clearable: boolean;
    optionsSourceAssociationCustomContentType: OptionsSourceAssociationCustomContentTypeEnum;
    optionsSourceAssociationCustomContent?: ListWidgetValue;
    optionsSourceDatabaseCustomContentType: OptionsSourceDatabaseCustomContentTypeEnum;
    optionsSourceDatabaseCustomContent?: ListWidgetValue;
    staticDataSourceCustomContentType: StaticDataSourceCustomContentTypeEnum;
    showFooter: boolean;
    menuFooterContent?: ReactNode;
    selectionMethod: SelectionMethodEnum;
    selectedItemsStyle: SelectedItemsStyleEnum;
    selectAllButton: boolean;
    selectAllButtonCaption: DynamicValue<string>;
    customEditability: CustomEditabilityEnum;
    customEditabilityExpression: DynamicValue<boolean>;
    readOnlyStyle: ReadOnlyStyleEnum;
    onChangeEvent?: ActionValue;
    onEnterEvent?: ActionValue;
    onLeaveEvent?: ActionValue;
    onChangeFilterInputEvent?: ActionValue<{ filterInput: Option<string> }>;
    ariaRequired: DynamicValue<boolean>;
    ariaLabel?: DynamicValue<string>;
    clearButtonAriaLabel?: DynamicValue<string>;
    removeValueAriaLabel?: DynamicValue<string>;
    a11ySelectedValue?: DynamicValue<string>;
    a11yOptionsAvailable?: DynamicValue<string>;
    a11yInstructions?: DynamicValue<string>;
    lazyLoading: boolean;
    loadingType: LoadingTypeEnum;
    selectedItemsSorting: SelectedItemsSortingEnum;
    filterType: FilterTypeEnum;
    filterInputDebounceInterval: number;
}